refactor: calculatefunc to be calculate
