+++ axe-check.js
@@ -1,7 +1,7 @@
 import { AxePuppeteer } from 'axe-puppeteer';
 import puppeteer from 'puppeteer';
 import path from 'path';
-import fs from 'fs';
+import fs from 'fs';git
 import { fileURLToPath } from 'url';
 
 const __filename = fileURLToPath(import.meta.url);
